Update file homework 3 task 21. Change length type int to double.

+++ cSharp/homeWork3_2/HomeWork3_2.cs
@@ -25,7 +25,7 @@ int y2 = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine("Please enter z2: ");
 int z2 = Convert.ToInt32(Console.ReadLine());
 
-int length = (int)Math.Sqrt(Math.Pow(GetDelta(x1, x2), 2) + Math.Pow(GetDelta(y1, y2), 2) + Math.Pow(GetDelta(z1, z2), 2));
+double length = Math.Sqrt(Math.Pow(GetDelta(x1, x2), 2) + Math.Pow(GetDelta(y1, y2), 2) + Math.Pow(GetDelta(z1, z2), 2));
 
 Console.Write("Length = ");
 Console.Write(length);
